multi-route: Add client and dedicated socket handler for multi-route support
Rather than support a single stream per socket approach we can instead
use the same websocket for multiple concurrent data-streams.  This commit
does so by adding the concept of a `client` which essentially an I/O wrapper
around the websocket.  The client sets up three dedicated communication
channels, two dedicated to stream subscriptions or publications and another
for "control" messages, which can be though of as commands or queries to/from
the client/server (yes, that means the server can issue commands to clients
as well).

In addition this starts prepping for the next phase of interface development
to provide an MVP client interface via clojurescript.